reservations_to_bq/requirements.txt
google.cloud.bigquery